Moved metrics classes from canton to daml after rewrite
Before this change, a lot of the metrics classes were living in the canton repo.
Now, I moved them back into the canton repository.
