+++ sdk/canton/community/domain/src/main/scala/com/digitalasset/canton/domain/metrics/DomainMetrics.scala
@@ -5,14 +5,13 @@ package com.digitalasset.canton.domain.metrics
 
 import com.daml.metrics.api.MetricHandle.*
 import com.daml.metrics.api.noop.NoOpMetricsFactory
-import com.daml.metrics.api.{MetricInfo, MetricName, MetricQualification, MetricsContext}
+import com.daml.metrics.api.{MetricInfo, MetricName, MetricQualification, MetricsContext, HistogramInventory}
 import com.daml.metrics.grpc.{DamlGrpcServerMetrics, GrpcServerMetrics}
 import com.daml.metrics.{CacheMetrics, HealthMetrics}
 import com.digitalasset.canton.environment.BaseMetrics
 import com.digitalasset.canton.metrics.{
   DbStorageHistograms,
   DbStorageMetrics,
-  HistogramInventory,
   SequencerClientHistograms,
   SequencerClientMetrics,
 }
