+++ sdk/canton/community/domain/src/test/scala/com/digitalasset/canton/domain/metrics/DomainTestMetrics.scala
@@ -7,7 +7,7 @@ import com.daml.metrics.HealthMetrics
 import com.daml.metrics.api.MetricName
 import com.daml.metrics.api.noop.NoOpMetricsFactory
 import com.daml.metrics.grpc.DamlGrpcServerMetrics
-import com.digitalasset.canton.metrics.HistogramInventory
+import com.daml.metrics.api.HistogramInventory
 
 object SequencerTestMetrics
     extends SequencerMetrics(
