+++ sdk/canton/community/ledger/ledger-common/src/main/scala/com/digitalasset/canton/metrics/ExecutionMetrics.scala
@@ -5,8 +5,8 @@ package com.digitalasset.canton.metrics
 
 import com.daml.metrics.CacheMetrics
 import com.daml.metrics.api.MetricHandle.*
-import com.daml.metrics.api.{MetricInfo, MetricName, MetricQualification}
-import com.digitalasset.canton.metrics.HistogramInventory.Item
+import com.daml.metrics.api.{HistogramInventory, MetricInfo, MetricName, MetricQualification}
+import com.daml.metrics.api.HistogramInventory.Item
 
 class ExecutionHistograms(val prefix: MetricName)(implicit
     inventory: HistogramInventory
