+++ sdk/canton/community/participant/src/main/scala/com/digitalasset/canton/participant/metrics/CommitmentMetrics.scala
@@ -5,8 +5,8 @@ package com.digitalasset.canton.participant.metrics
 
 import com.daml.metrics.api.MetricHandle.{Gauge, LabeledMetricsFactory, Meter, Timer}
 import com.daml.metrics.api.{MetricInfo, MetricName, MetricQualification, MetricsContext}
-import com.digitalasset.canton.metrics.HistogramInventory
-import com.digitalasset.canton.metrics.HistogramInventory.Item
+import com.daml.metrics.api.HistogramInventory
+import com.daml.metrics.api.HistogramInventory.Item
 
 class CommitmentHistograms(parent: MetricName)(implicit inventory: HistogramInventory) {
   private[metrics] val prefix = parent :+ "commitments"
