+++ sdk/canton/community/participant/src/test/scala/com/digitalasset/canton/participant/metrics/ParticipantTestMetrics.scala
@@ -6,7 +6,7 @@ package com.digitalasset.canton.participant.metrics
 import com.daml.metrics.api.MetricName
 import com.daml.metrics.api.noop.NoOpMetricsFactory
 import com.digitalasset.canton.DomainAlias
-import com.digitalasset.canton.metrics.HistogramInventory
+import com.daml.metrics.api.HistogramInventory
 
 object ParticipantTestMetrics
     extends ParticipantMetrics(
