+++ sdk/observability/pekko-http-metrics/src/main/scala/com/daml/metrics/http/DamlWebSocketMetrics.scala
@@ -6,11 +6,36 @@ package com.daml.metrics.http
 import com.daml.metrics.api.MetricQualification
 import com.daml.metrics.api.MetricHandle.{Histogram, LabeledMetricsFactory, Meter}
 import com.daml.metrics.api.{MetricInfo, MetricName, MetricsContext}
+import com.daml.metrics.api.HistogramInventory
+import com.daml.metrics.api.HistogramInventory.Item
 
-class DamlWebSocketMetrics(metricsFactory: LabeledMetricsFactory, component: String)
-    extends WebSocketMetrics {
+class DamlWebSocketsHistograms(implicit
+    inventory: HistogramInventory
+) {
+  private[http] val websocketMetricsPrefix = MetricName.Daml :+ "http" :+ "websocket" :+ "messages"
 
-  private val httpMetricsPrefix = MetricName.Daml :+ "http"
+  val messagesReceivedBytes: Item =
+    Item(
+      websocketMetricsPrefix :+ "received" :+ Histogram.Bytes,
+      "Distribution of the size of received WebSocket messages.",
+      MetricQualification.Debug,
+    )
+
+  val messagesSentBytes: Item =
+    Item(
+      websocketMetricsPrefix :+ "sent" :+ Histogram.Bytes,
+      "Distribution of the size of sent WebSocket messages.",
+      MetricQualification.Debug,
+    )
+
+}
+
+class DamlWebSocketMetrics(
+    metricsFactory: LabeledMetricsFactory,
+    component: String,
+) extends WebSocketMetrics {
+
+  private val histograms = new DamlWebSocketsHistograms()(new HistogramInventory)
 
   private implicit val metricsContext: MetricsContext = MetricsContext(
     Map(Labels.ServiceLabel -> component)
@@ -19,34 +44,23 @@ class DamlWebSocketMetrics(metricsFactory: LabeledMetricsFactory, component: Str
   override val messagesReceivedTotal: Meter =
     metricsFactory.meter(
       MetricInfo(
-        httpMetricsPrefix :+ "websocket" :+ "messages" :+ "received",
+        histograms.websocketMetricsPrefix :+ "received",
         "Total number of received WebSocket messages.",
         MetricQualification.Debug,
       )
     )
   override val messagesReceivedBytes: Histogram =
-    metricsFactory.histogram(
-      MetricInfo(
-        httpMetricsPrefix :+ "websocket" :+ "messages" :+ "received" :+ Histogram.Bytes,
-        "Distribution of the size of received WebSocket messages.",
-        MetricQualification.Debug,
-      )
-    )
+    metricsFactory.histogram(histograms.messagesReceivedBytes.info)
+
   override val messagesSentTotal: Meter =
     metricsFactory.meter(
       MetricInfo(
-        httpMetricsPrefix :+ "websocket" :+ "messages" :+ "sent",
+        histograms.websocketMetricsPrefix :+ "sent",
         "Total number of sent WebSocket messages.",
         MetricQualification.Debug,
       )
     )
   override val messagesSentBytes: Histogram =
-    metricsFactory.histogram(
-      MetricInfo(
-        httpMetricsPrefix :+ "websocket" :+ "messages" :+ "sent" :+ Histogram.Bytes,
-        "Distribution of the size of sent WebSocket messages.",
-        MetricQualification.Debug,
-      )
-    )
+    metricsFactory.histogram(histograms.messagesSentBytes.info)
 
 }
