--- sdk/observability/telemetry/BUILD.bazel
@@ -1,24 +1,0 @@
-# Copyright (c) 2024 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-load(
-    "//bazel_tools:scala.bzl",
-    "da_scala_library",
-)
-
-da_scala_library(
-    name = "telemetry",
-    srcs = glob(["src/main/scala/**/*.scala"]),
-    resources = glob(["src/main/resources/**/*"]),
-    scala_deps = [
-    ],
-    tags = ["maven_coordinates=com.daml:telemetry:__VERSION__"],
-    visibility = [
-        "//visibility:public",
-    ],
-    runtime_deps = [],
-    deps = [
-        "//observability/metrics",
-        "@maven//:io_opentelemetry_opentelemetry_sdk_metrics",
-    ],
-)
