+++ sdk/release/artifacts.yaml
@@ -149,8 +149,6 @@
   type: jar-scala
 - target: //observability/tracing:tracing-test-lib
   type: jar-scala
-- target: //observability/telemetry:telemetry
-  type: jar-scala
 - target: //observability/pekko-http-metrics:pekko-http-metrics
   type: jar-scala
 - target: //canton:community_ledger_ledger-common
